pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
readme = "README.md"
license = { file = "LICENSE" }
name = "django-import-export-gsheets"
version = "0.3.0"
description = "Google Sheets Tablib format for django-import-export with django-allauth"
requires-python = ">=3.9"
dependencies = [
  "Django>=3.2",
  "django-import-export>=3.0",
  "tablib>=3.5",
  "django-allauth>=0.61",
  "google-auth>=2.0",
  "google-auth-oauthlib>=1.0",
  "google-api-python-client>=2.0",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
